#17: DownloadListSerializerImpl split onto base trait and implementation
